Alias == as eql?
Similar to what ActiveRecord does, add an alias to the :== method for eql?. This allows a StoreModel to be used as a hash key or in set operations. Note that we use alias here and not alias_method since this method is defined in a module. All of the related specs are extracted into a shared example that is now used for all comparison methods